Refactor project into kyo_tool package
