--- ai_extractor.py
@@ -1,8 +1,0 @@
-from extract.ai_extractor import QAExtractor
-
-_extractor = QAExtractor()
-
-ai_extract = _extractor.extract
-extract_qa_numbers = _extractor.extract_qa_numbers
-extract_models = _extractor.extract_models
-extract_dates = _extractor.extract_dates
