+++ kyo_tool/__init__.py
@@ -1,0 +1,5 @@
+"""Top-level package for the KYO QA Knowledge Tool."""
+
+from .version import VERSION
+
+__all__ = ["VERSION"]
