+++ kyo_tool/transform/__init__.py
@@ -1,7 +1,12 @@
 # KYO QA ServiceNow Data Harvesters
 import re
-from logging_utils import setup_logger, log_info, log_error, log_warning
-from config import (
+from kyo_tool.utils.logging_utils import (
+    setup_logger,
+    log_info,
+    log_error,
+    log_warning,
+)
+from kyo_tool.utils.config import (
     DATE_PATTERNS,
     SUBJECT_PATTERNS,
     STANDARDIZATION_RULES,
@@ -68,7 +73,7 @@ def harvest_subject(text):
 def harvest_metadata(text, pdf_path=None):
     """Extract comprehensive metadata from document text and PDF metadata."""
     try:
-        from ocr_utils import get_pdf_metadata
+        from kyo_tool.utils.ocr_utils import get_pdf_metadata
         pdf_metadata = get_pdf_metadata(pdf_path) if pdf_path else {}
         results = {}
