+++ kyo_tool/utils/custom_exceptions.py
