+++ kyo_tool/utils/ocr_utils.py
@@ -3,7 +3,12 @@
 import fitz  # PyMuPDF
 import os
 from pathlib import Path
-from logging_utils import setup_logger, log_info, log_error, log_warning
+from kyo_tool.utils.logging_utils import (
+    setup_logger,
+    log_info,
+    log_error,
+    log_warning,
+)
 
 logger = setup_logger("ocr_utils")
